Start of Camera Scripting
Camera script seems to be a lot more complicated than originally thought. Basic jist is that I need to make the player move in the World space, not Local, and apply rotation to the player using their movement.

Movement will need to be based off an origin, which will be a standard Vector3.forward, but can be replaced with the forward direction relative to the rotation of the camera. As the camera moves and rotates independantly (except in certain situations) this shouldn't lead to unexpected rotations.