Sprint Attack, framework for Dash & Dash Attack
Added in the ability to use an attack whilst sprinting, including appropriate state tracking and speed.

Added framework for dashing, and a basic implementation of a dash. The framework also exists for Dash Attacks.

Changed the rotation whilst attacking to be a LookAt again, I feel this is better as it provides more control to the player. During testing an issue was bought to light regarding the second/third animation cancelling during a light attack combo when spamming clicks, and the AttackNumber variable going on far beyond what it should be. Likely introduced as part of the dashing as I havent seen it until now.