Fixed attacking, Sprint attack and finsihed Dash & Dash Attack
Fixed the issue with skipping attacks and animations, wasn't fully tracking state and certain actions could be taken repeatedly that shouldnt have been.

Sprint attacking has been fixed, with a small delay after the animation to allow the player to be vulerable afterwards.

Dashing has been implemented properly, and there is a slight delay after dashing has finished for balancing purposes. A dash can also be followed up by an attack, which can only be performed in the direction the player is dashing in.