Lots of Stuff
Added basic lock-on, including rotation towards enemy during attacks.

Added "death" and respawning, added a killzone to compliment this.

Changed the basic plane and boxes to a custom terrain. The terrain is still flat on top, this will be worked on when aerial attacks are implemented and knockback "bouncing" is added.