Whitebox level layout and Trigger script changes
Created a trigger class, containing a nested class (triggerStruct, which isnt a struct but was one originally, a class fit better) which contains all information needed to fire an event, and a method to do so. The trigger class has a list of these subclasses which can be added to or removed from at will. The subclasses have an accessible parameters property which allows the user to set parameters without using the constructor. This can overwrite existing parameters for already set triggers, or can add params to ones created through the inspector.

Added ability to add triggers through the inspector, including a custom Editor list. They still need to be called via code and they cannot have parameters (unless set later via code), but otherwise can be completely configured via the inspector.